+++ pyodide-build/pyodide_build/pywasmcross.py
@@ -21,12 +21,11 @@ if IS_MAIN:
 
     PYWASMCROSS_ARGS["pythoninclude"] = os.environ["PYTHONINCLUDE"]
 
-import importlib.machinery
 import re
 import subprocess
 from collections import namedtuple
 from pathlib import Path, PurePosixPath
-from typing import Any, MutableMapping, NoReturn, Optional, overload
+from typing import Any, MutableMapping, NoReturn, overload
 
 # absolute import is necessary as this file will be symlinked
 # under tools
@@ -123,7 +122,7 @@ def compile(env, **kwargs):
         raise
 
 
-def replay_f2c(args: list[str], dryrun: bool = False) -> Optional[list[str]]:
+def replay_f2c(args: list[str], dryrun: bool = False) -> list[str] | None:
     """Apply f2c to compilation arguments
 
     Parameters
@@ -171,7 +170,7 @@ def replay_f2c(args: list[str], dryrun: bool = False) -> Optional[list[str]]:
     return new_args
 
 
-def get_library_output(line: list[str]) -> Optional[str]:
+def get_library_output(line: list[str]) -> str | None:
     """
     Check if the command is a linker invocation. If so, return the name of the
     output file.
@@ -210,7 +209,7 @@ def parse_replace_libs(replace_libs: str) -> dict[str, str]:
 
 def replay_genargs_handle_dashl(
     arg: str, replace_libs: dict[str, str], used_libs: set[str]
-) -> Optional[str]:
+) -> str | None:
     """
     Figure out how to replace a `-lsomelib` argument.
 
@@ -251,7 +250,7 @@ def replay_genargs_handle_dashl(
     return arg
 
 
-def replay_genargs_handle_dashI(arg: str, target_install_dir: str) -> Optional[str]:
+def replay_genargs_handle_dashI(arg: str, target_install_dir: str) -> str | None:
     """
     Figure out how to replace a `-Iincludepath` argument.
 
@@ -313,7 +312,7 @@ def replay_genargs_handle_linker_opts(arg):
         return None
 
 
-def replay_genargs_handle_argument(arg: str) -> Optional[str]:
+def replay_genargs_handle_argument(arg: str) -> str | None:
     """
     Figure out how to replace a general argument.
 
@@ -502,8 +501,7 @@ def handle_command(
        containing ``args.cflags``, ``args.cxxflags``, and ``args.ldflags``
     """
     # some libraries have different names on wasm e.g. png16 = png
-    library_output = get_library_output(line)
-    is_link_cmd = library_output is not None
+    is_link_cmd = get_library_output(line) is not None
 
     if line[0] == "gfortran":
         if "-dumpversion" in line:
@@ -522,22 +520,11 @@ def handle_command(
     if returncode != 0:
         sys.exit(returncode)
 
-    # Emscripten .so files shouldn't have the native platform slug
-    if library_output:
-        renamed = library_output
-        for ext in importlib.machinery.EXTENSION_SUFFIXES:
-            if ext == ".so":
-                continue
-            if renamed.endswith(ext):
-                renamed = renamed[: -len(ext)] + ".so"
-                break
-        if library_output != renamed:
-            os.rename(library_output, renamed)
     sys.exit(returncode)
 
 
 def environment_substitute_args(
-    args: dict[str, str], env: dict[str, str] = None
+    args: dict[str, str], env: dict[str, str] | None = None
 ) -> dict[str, Any]:
     if env is None:
         env = dict(os.environ)