+++ src/py/_pyodide/_base.py
@@ -11,7 +11,7 @@ from copy import deepcopy
 from io import StringIO
 from textwrap import dedent
 from types import CodeType
-from typing import Any, Generator, Optional
+from typing import Any, Generator
 
 
 def should_quiet(source: str) -> bool:
@@ -254,7 +254,11 @@ class CodeRunner:
             raise AssertionError()
         return self
 
-    def run(self, globals: dict[str, Any] = None, locals: dict[str, Any] = None):
+    def run(
+        self,
+        globals: dict[str, Any] | None = None,
+        locals: dict[str, Any] | None = None,
+    ):
         """Executes ``self.code``.
 
         Can only be used after calling compile. The code may not use top level
@@ -303,7 +307,9 @@ class CodeRunner:
             return e.value
 
     async def run_async(
-        self, globals: dict[str, Any] = None, locals: dict[str, Any] = None
+        self,
+        globals: dict[str, Any] | None = None,
+        locals: dict[str, Any] | None = None,
     ):
         """Runs ``self.code`` which may use top level await.
 
@@ -351,8 +357,8 @@ class CodeRunner:
 
 def eval_code(
     source: str,
-    globals: Optional[dict[str, Any]] = None,
-    locals: Optional[dict[str, Any]] = None,
+    globals: dict[str, Any] | None = None,
+    locals: dict[str, Any] | None = None,
     *,
     return_mode: str = "last_expr",
     quiet_trailing_semicolon: bool = True,
@@ -426,8 +432,8 @@ def eval_code(
 
 async def eval_code_async(
     source: str,
-    globals: Optional[dict[str, Any]] = None,
-    locals: Optional[dict[str, Any]] = None,
+    globals: dict[str, Any] | None = None,
+    locals: dict[str, Any] | None = None,
     *,
     return_mode: str = "last_expr",
     quiet_trailing_semicolon: bool = True,